+++ common/src/main/java/mod/azure/azurelib/rewrite/render/AzRendererConfig.java
@@ -36,9 +36,11 @@ public class AzRendererConfig<T> {
 
     private final Function<T, ResourceLocation> textureLocationProvider;
 
-    private final float scaleHeight;
+    private final Function<T, Float> alphaFunction;
 
-    private final float scaleWidth;
+    private final Function<T, Float> scaleHeight;
+
+    private final Function<T, Float> scaleWidth;
 
     public AzRendererConfig(
         Supplier<AzAnimator<T>> animatorProvider,
@@ -48,8 +50,9 @@ public class AzRendererConfig<T> {
         Function<AzRendererPipelineContext<T>, AzRendererPipelineContext<T>> preRenderEntry,
         Function<AzRendererPipelineContext<T>, AzRendererPipelineContext<T>> postRenderEntry,
         Function<T, ResourceLocation> textureLocationProvider,
-        float scaleHeight,
-        float scaleWidth
+        Function<T, Float> alphaFunction,
+        Function<T, Float> scaleHeight,
+        Function<T, Float> scaleWidth
     ) {
         this.animatorProvider = animatorProvider;
         this.modelLocationProvider = modelLocationProvider;
@@ -58,6 +61,7 @@ public class AzRendererConfig<T> {
         this.preRenderEntry = preRenderEntry;
         this.postRenderEntry = postRenderEntry;
         this.textureLocationProvider = textureLocationProvider;
+        this.alphaFunction = alphaFunction;
         this.scaleHeight = scaleHeight;
         this.scaleWidth = scaleWidth;
     }
@@ -90,12 +94,16 @@ public class AzRendererConfig<T> {
         return postRenderEntry.apply(animatable);
     }
 
-    public float scaleHeight() {
-        return scaleHeight;
+    public float alpha(T entity) {
+        return alphaFunction.apply(entity);
     }
 
-    public float scaleWidth() {
-        return scaleWidth;
+    public float scaleHeight(T entity) {
+        return scaleHeight.apply(entity);
+    }
+
+    public float scaleWidth(T entity) {
+        return scaleWidth.apply(entity);
     }
 
     public static class Builder<T> {
@@ -114,9 +122,11 @@ public class AzRendererConfig<T> {
 
         protected Supplier<@Nullable AzAnimator<T>> animatorProvider;
 
-        protected float scaleHeight;
+        protected Function<T, Float> alphaFunction;
 
-        protected float scaleWidth;
+        protected Function<T, Float> scaleHeight;
+
+        protected Function<T, Float> scaleWidth;
 
         protected Builder(
             Function<T, ResourceLocation> modelLocationProvider,
@@ -129,8 +139,9 @@ public class AzRendererConfig<T> {
             this.preRenderEntry = $ -> $;
             this.postRenderEntry = $ -> $;
             this.textureLocationProvider = textureLocationProvider;
-            this.scaleHeight = 1;
-            this.scaleWidth = 1;
+            this.alphaFunction = $ -> 1.0F;
+            this.scaleHeight = $ -> 1.0F;
+            this.scaleWidth = $ -> 1.0F;
         }
 
         public Builder<T> setPrerenderEntry(
@@ -169,6 +180,32 @@ public class AzRendererConfig<T> {
         }
 
         /**
+         * Sets the alpha value provider for the builder. The alpha value determines the opacity level of the rendered
+         * object and is calculated dynamically based on the specified function.
+         *
+         * @param alphaFunction a {@link Function} that takes an object of type {@code T} and returns a {@code Float}
+         *                      value representing the alpha (opacity) level, where 0.0 is fully transparent and 1.0 is
+         *                      fully opaque
+         * @return the updated {@code Builder} instance for chaining configuration methods
+         */
+        public Builder<T> setAlpha(Function<T, Float> alphaFunction) {
+            this.alphaFunction = alphaFunction;
+            return this;
+        }
+
+        /**
+         * Sets the alpha transparency level for the builder, which determines the level of transparency to be applied.
+         *
+         * @param alpha the alpha transparency value to set, where 0.0 represents fully transparent and 1.0 represents
+         *              fully opaque
+         * @return the updated {@code Builder} instance for chaining configuration methods
+         */
+        public Builder<T> setAlpha(float alpha) {
+            this.alphaFunction = $ -> alpha;
+            return this;
+        }
+
+        /**
          * Sets the scaling factor uniformly for both width and height dimensions.
          *
          * @param scale the uniform scaling factor to be applied to both width and height
@@ -186,8 +223,38 @@ public class AzRendererConfig<T> {
          * @return the updated builder instance for chaining operations
          */
         public Builder<T> setScale(float scaleWidth, float scaleHeight) {
-            this.scaleHeight = scaleHeight;
-            this.scaleWidth = scaleWidth;
+            this.scaleHeight = $ -> scaleHeight;
+            this.scaleWidth = $ -> scaleWidth;
+            return this;
+        }
+
+        /**
+         * Sets the scaling function for both the width and height dimensions of the target object. The provided
+         * function dynamically calculates scaling factors based on the input object of type {@code T}.
+         *
+         * @param scaleFunction a {@link Function} that takes an object of type {@code T} and returns a {@code Float}
+         *                      value representing the scaling factor to be applied uniformly to both width and height
+         * @return the updated {@code Builder} instance for chaining configuration methods
+         */
+        public Builder<T> setScale(Function<T, Float> scaleFunction) {
+            this.scaleHeight = scaleFunction;
+            this.scaleWidth = scaleFunction;
+            return this;
+        }
+
+        /**
+         * Sets the scaling functions for height and width dimensions. These functions dynamically calculate scaling
+         * factors based on the input object of type {@code T}.
+         *
+         * @param scaleHeightFunction a {@link Function} that takes an object of type {@code T} and returns a
+         *                            {@code Float} representing the scaling factor for the height dimension
+         * @param scaleWidthFunction  a {@link Function} that takes an object of type {@code T} and returns a
+         *                            {@code Float} representing the scaling factor for the width dimension
+         * @return the updated {@code Builder} instance for chaining configuration methods
+         */
+        public Builder<T> setScale(Function<T, Float> scaleHeightFunction, Function<T, Float> scaleWidthFunction) {
+            this.scaleHeight = scaleHeightFunction;
+            this.scaleWidth = scaleWidthFunction;
             return this;
         }
 
@@ -207,6 +274,7 @@ public class AzRendererConfig<T> {
                 preRenderEntry,
                 postRenderEntry,
                 textureLocationProvider,
+                alphaFunction,
                 scaleHeight,
                 scaleWidth
             );