+++ common/src/main/java/mod/azure/azurelib/rewrite/render/block/AzBlockEntityRendererConfig.java
@@ -31,8 +31,9 @@ public class AzBlockEntityRendererConfig<T extends BlockEntity> extends AzRender
         Function<AzRendererPipelineContext<T>, AzRendererPipelineContext<T>> preRenderEntry,
         Function<AzRendererPipelineContext<T>, AzRendererPipelineContext<T>> postRenderEntry,
         Function<T, ResourceLocation> textureLocationProvider,
-        float scaleHeight,
-        float scaleWidth
+        Function<T, Float> alphaFunction,
+        Function<T, Float> scaleHeight,
+        Function<T, Float> scaleWidth
     ) {
         super(
             animatorProvider,
@@ -42,6 +43,7 @@ public class AzBlockEntityRendererConfig<T extends BlockEntity> extends AzRender
             preRenderEntry,
             postRenderEntry,
             textureLocationProvider,
+            alphaFunction,
             scaleHeight,
             scaleWidth
         );
@@ -105,6 +107,36 @@ public class AzBlockEntityRendererConfig<T extends BlockEntity> extends AzRender
         }
 
         @Override
+        public Builder<T> setAlpha(Function<T, Float> alphaFunction) {
+            return (AzBlockEntityRendererConfig.Builder<T>) super.setAlpha(alphaFunction);
+        }
+
+        @Override
+        public Builder<T> setAlpha(float alpha) {
+            return (AzBlockEntityRendererConfig.Builder<T>) super.setAlpha(alpha);
+        }
+
+        @Override
+        public Builder<T> setScale(Function<T, Float> scaleFunction) {
+            return (AzBlockEntityRendererConfig.Builder) super.setScale(scaleFunction);
+        }
+
+        @Override
+        public Builder<T> setScale(Function<T, Float> scaleHeightFunction, Function<T, Float> scaleWidthFunction) {
+            return (AzBlockEntityRendererConfig.Builder) super.setScale(scaleHeightFunction, scaleWidthFunction);
+        }
+
+        @Override
+        public Builder<T> setScale(float scale) {
+            return (AzBlockEntityRendererConfig.Builder<T>) super.setScale(scale);
+        }
+
+        @Override
+        public Builder<T> setScale(float scaleWidth, float scaleHeight) {
+            return (AzBlockEntityRendererConfig.Builder<T>) super.setScale(scaleWidth, scaleHeight);
+        }
+
+        @Override
         public AzBlockEntityRendererConfig<T> build() {
             var baseConfig = super.build();
 
@@ -116,8 +148,9 @@ public class AzBlockEntityRendererConfig<T extends BlockEntity> extends AzRender
                 baseConfig::preRenderEntry,
                 baseConfig::postRenderEntry,
                 baseConfig::textureLocation,
-                baseConfig.scaleHeight(),
-                baseConfig.scaleWidth()
+                baseConfig::alpha,
+                baseConfig::scaleHeight,
+                baseConfig::scaleWidth
             );
         }
     }