+++ common/src/main/java/mod/azure/azurelib/rewrite/render/entity/AzEntityRendererPipeline.java
@@ -71,10 +71,15 @@ public class AzEntityRendererPipeline<T extends Entity> extends AzRendererPipeli
         this.entityRenderTranslations.set(poseStack.last().pose());
 
         var config = entityRenderer.config();
-        var scaleWidth = config.scaleWidth();
-        var scaleHeight = config.scaleHeight();
+        var scaleWidth = config.scaleWidth(context.animatable());
+        var scaleHeight = config.scaleHeight(context.animatable());
 
         scaleModelForRender(context, scaleWidth, scaleHeight, isReRender);
+        if (config.alpha(context.animatable()) < 1) {
+            var alpha = (int) (config.alpha(context.animatable()) * 0xFF) << 24;
+            var color = (context.renderColor() & 0xFFFFFF) | alpha;
+            context.setRenderColor(color);
+        }
         config.preRenderEntry(context);
     }