+++ common/src/main/java/mod/azure/azurelib/rewrite/render/item/AzItemRendererConfig.java
@@ -33,8 +33,9 @@ public class AzItemRendererConfig extends AzRendererConfig<ItemStack> {
         Function<AzRendererPipelineContext<ItemStack>, AzRendererPipelineContext<ItemStack>> preRenderEntry,
         Function<AzRendererPipelineContext<ItemStack>, AzRendererPipelineContext<ItemStack>> postRenderEntry,
         Function<ItemStack, ResourceLocation> textureLocationProvider,
-        float scaleHeight,
-        float scaleWidth,
+        Function<ItemStack, Float> alphaFunction,
+        Function<ItemStack, Float> scaleHeight,
+        Function<ItemStack, Float> scaleWidth,
         boolean useEntityGuiLighting,
         boolean useNewOffset
     ) {
@@ -46,6 +47,7 @@ public class AzItemRendererConfig extends AzRendererConfig<ItemStack> {
             preRenderEntry,
             postRenderEntry,
             textureLocationProvider,
+            alphaFunction,
             scaleHeight,
             scaleWidth
         );
@@ -130,6 +132,36 @@ public class AzItemRendererConfig extends AzRendererConfig<ItemStack> {
             return this;
         }
 
+        @Override
+        public Builder setAlpha(Function<ItemStack, Float> alphaFunction) {
+            return (AzItemRendererConfig.Builder) super.setAlpha(alphaFunction);
+        }
+
+        @Override
+        public Builder setAlpha(float alpha) {
+            return (AzItemRendererConfig.Builder) super.setAlpha(alpha);
+        }
+
+        @Override
+        public Builder setScale(Function<ItemStack, Float> scaleFunction) {
+            return (AzItemRendererConfig.Builder) super.setScale(scaleFunction);
+        }
+
+        @Override
+        public Builder setScale(Function<ItemStack, Float> scaleHeightFunction, Function<ItemStack, Float> scaleWidthFunction) {
+            return (AzItemRendererConfig.Builder) super.setScale(scaleHeightFunction, scaleWidthFunction);
+        }
+
+        @Override
+        public Builder setScale(float scale) {
+            return (AzItemRendererConfig.Builder) super.setScale(scale);
+        }
+
+        @Override
+        public Builder setScale(float scaleWidth, float scaleHeight) {
+            return (AzItemRendererConfig.Builder) super.setScale(scaleWidth, scaleHeight);
+        }
+
         /**
          * @param useNewOffset Determines whether to apply the y offset for a model due to the change in BlockBench
          *                     4.11.
@@ -151,8 +183,9 @@ public class AzItemRendererConfig extends AzRendererConfig<ItemStack> {
                 baseConfig::preRenderEntry,
                 baseConfig::postRenderEntry,
                 baseConfig::textureLocation,
-                baseConfig.scaleHeight(),
-                baseConfig.scaleWidth(),
+                baseConfig::alpha,
+                baseConfig::scaleHeight,
+                baseConfig::scaleWidth,
                 useEntityGuiLighting,
                 useNewOffset
             );