+++ common/src/main/java/mod/azure/azurelib/rewrite/render/item/AzItemRendererPipeline.java
@@ -51,17 +51,24 @@ public class AzItemRendererPipeline extends AzRendererPipeline<ItemStack> {
      */
     @Override
     public void preRender(AzRendererPipelineContext<ItemStack> context, boolean isReRender) {
-        var poseStack = context.poseStack();
+        var itemContext = (AzItemRendererPipelineContext) context;
+        var poseStack = itemContext.poseStack();
         this.itemRenderTranslations = new Matrix4f(poseStack.last().pose());
 
         var config = itemRenderer.config();
-        var scaleWidth = config.scaleWidth();
-        var scaleHeight = config.scaleHeight();
-        scaleModelForRender(context, scaleWidth, scaleHeight, isReRender);
+        var scaleWidth = config.scaleWidth(context.animatable());
+        var scaleHeight = config.scaleHeight(context.animatable());
+        scaleModelForRender(itemContext, scaleWidth, scaleHeight, isReRender);
 
         if (!isReRender) {
             var useNewOffset = config.useNewOffset();
             poseStack.translate(0.5f, useNewOffset ? 0.0f : 0.51f, 0.5f);
+        }
+        if (config.alpha(context.animatable()) < 1) {
+            var alpha = (int) (config.alpha(context.animatable()) * 0xFF) << 24;
+            var color = (itemContext.renderColor() & 0xFFFFFF) | alpha;
+            itemContext.setRenderColor(color);
+            itemContext.setTranslucent(true);
         }
         config.preRenderEntry(context);
     }