+++ changelog.md
@@ -1,3 +1,5 @@
-v3.0.16
+v3.0.17
 
-- Adds Shoulder Surfing support for armor.
+- Adds builder support for Alpha, fixing shader issue from previous update. Thanks daedelus_dev for the fix!
+- Fixes scale not working for builders.
+- Adds scale builders with context of the entity being rendered.